+++ CHANGELOG.rst
@@ -3,6 +3,8 @@
 
 ### v1.1.0 (unreleased)
 
+- Upgrade scancode-toolkit to version 21.2.9
+
 - Allow to start the pipeline run immediately on addition in the `add_pipeline` action
   of the Project API endpoint.
   https://github.com/nexB/scancode.io/issues/92