+++ CHANGELOG.rst
@@ -4,6 +4,7 @@
 ### v1.1.0 (unreleased)
 
 - Add a "Run Pipeline" button in the Pipeline modal of the Project details view
+  Pipelines can now be added from the Project details view
   https://github.com/nexB/scancode.io/issues/84
 
 - Upgrade scancode-toolkit to version 21.2.9